% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodeplot_colorless.R
\name{barcodeplot_colorless}
\alias{barcodeplot_colorless}
\title{Generate a barcodeplot without the colors on the axis}
\usage{
\code{barcodeplot_colorless(statistics, index = NULL, index2 = NULL, gene.weights = NULL,
 weights.label = "Weight", labels = c("Up", "Down"),
 quantiles = c(-1,1) * sqrt(2), col.bars = NULL, alpha = 0.4, worm = TRUE,
 span.worm = 0.45, ...)}
}
\value{
A data frame with a single row, containing the p-value, the upper and lower confidence intervals, and the estimated odds ratio.
}
\description{
Plot the gene statistics of a set of genes on a barcode plot. This is a modified version of
limma::barcodeplot that drops the pink and blue colors in the limma version. For documentation, see
limma::barcodeplot.
}
