% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deduplicate_list_elements.R
\name{deduplicate_list_elements}
\alias{deduplicate_list_elements}
\title{Remove duplicate elements from a list of vectors}
\usage{
\code{deduplicate_list_elements(data, remove="last")}
}
\arguments{
\item{data}{a list of character vectors.}

\item{remove}{which duplicate to remove; "first" always removes the first element found in another vector, "last" removes the last, and "random" selects one at random to remove}
}
\value{
A data frame with a row for each gene set, containing the name of the gene set, the p-value, the upper and lower confidence intervals, and the estimated odds ratio.
}
\description{
Remove duplicate elements from a list of vectors. This cycles through a list of vectors, repeatedly
checking the elements of each vector against all other vectors in the list. If any elements are
shared with other vectors, one is removed (can be the first or last such duplicate in the vector),
and the algorithm moves to the next vector. This continues until no elements are included in more than
one of the vectors in the list. It is used in the context of gene set analyses to remove genes that
are duplicated across multiple gene sets, so that the gene sets are non-overlapping.
}

