% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gene_sets.R
\name{filter_gene_sets}
\alias{filter_gene_sets}
\title{Filter gene sets}
\usage{
\code{filter_gene_sets(
             gene_sets, counts,
             counts_genes_by="rows",
             remove_missing_genes=TRUE,
             remove_low_count_genes=FALSE,
             min_median_gene_expression=1,
             min_genes=2,
             min_median_gene_set_expression=1)}
}
\arguments{
\item{gene_sets}{the gene set(s). Can be one of (1) a single gene set, as character vector, (2) a list of gene sets, with each as a character vector, or (3) a matrix or data frame, with one gene set per column, and the gene set names as column names. Names of list elements or columns are used as gene set names.}

\item{counts}{a matrix or data frame of counts. The dimension corresponding to genes must have dimnames, which will be matched to the gene identifiers in \code{gene_sets}.}

\item{counts_genes_by}{the dimension of \code{counts} corresponding to genes. Can be "rows" or "columns" or partial matches.}

\item{remove_missing_genes}{logical, whether to remove genes from the gene sets if the genes are not present in the counts object. Defaults to TRUE.}

\item{remove_low_count_genes}{logical, whether to remove genes from the gene sets if their median count does not exceed a threshold value. Defaults to FALSE.}

\item{min_median_gene_expression}{numeric, the minimum value for median expression of genes. Any genes with median expression lower than this value are removed from gene sets. Defaults to 1. Ignored if \code{remove_low_count_genes} is FALSE.}

\item{min_genes}{an integer, the minimum number of genes from a gene set that must be in the data. Gene sets with fewer genes overlapping with the dataset are dropped. Defaults to 2.}

\item{min_median_gene_set_expression}{numeric, the minimum value for median expression of genes in a gene set. Any gene sets with median expression lower than this value are removed.}
}
\value{
An object of the same class as the input object \code{gene_sets}, with the gene sets that meet the thresholds.
}
\description{
This function filters one or more gene sets based on a set of criteria, removing gene sets that do not meet those criteria. The standard criterion is median gene expression above a certain threshold, but other functionality may be built in at a later time. It can optionally remove genes not present in the counts object, and/or remove genes from gene sets if their median counts are below a certain threshold.
}
\seealso{
\code{\link{gene_set_expressed}}
}

