% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_enrich_terms.R
\name{gene_set_enrich_terms}
\alias{gene_set_enrich_terms}
\title{Extract enrichment terms from a list of gene sets significantly up and/or down in a gene set analysis}
\usage{
\code{
gene_set_enrich_terms(
  results, gene_set_df, method=NULL, 
  results.name_col=NULL, results.sig_col=NULL, results.direction_col=NULL,
  gene_set_df.name_col, gene_set_df.terms_col,
  threshold=0.01, direction=NULL)}
}
\arguments{
\item{results}{the output from a gene set analysis (e.g. roast, mroast, camera, gene)}

\item{gene_set_df}{a data frame containing the details of the gene set annotations.}

\item{method}{a string, the method used to run the gene set analysis. This provides the function with the expected structure of \code{results}, so that various other parameters do not need to be specified. Current options include "camera", "mroast", "gene_set_list_exact_test", and "WGCNA_modules_gene_set_list_exact_test}

\item{results.name_col}{name or number of the column of \code{results} containing the gene set names. If gene set names are the row names, specify "rownames".}

\item{results.sig_col}{name or number of the column of \code{results} containing the gene set significance numbers.}

\item{results.direction_col}{name number of the column of \code{results} containing the direction (Up/Down).}

\item{gene_set_df.name_col}{name or number of the column of \code{gene_set_df} containing the gene set names. If gene set names are the row names, specify "rownames".}

\item{gene_set_df.terms_col}{name or number of the column of \code{gene_set_df} containing the gene set terms or descriptions.}

\item{threshold}{numeric, the maximum value of the significance statistic for which to include gene sets. Defaults to 0.01.}

\item{direction}{character, the direction of gene sets to include. Can be "up", "down", or "both". Only used if results object contains a column for direction.}
}
\value{
A data frame (or list of data frames, one for each direction) containing the names of the enriched gene sets and the terms or description associated with each gene set.
}
\description{
This is a fetch function; it generates vectors of terms associated with gene sets that are
significantly up or down in a gene set analysis. Basically, it takes the output of the gene set
analysis, identifies gene sets that are significantly up and/or down in the that analysis, and
extracts terms related to those gene sets from a data frame describing the gene sets.
}
