% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_expressed.R
\name{gene_set_expressed}
\alias{gene_set_expressed}
\title{Trim low-count genes from gene sets}
\usage{
\code{gene_set_expressed(gene_set, counts,
  counts_genes_by="rows",
  remove_missing_genes=TRUE,
  min_median_gene_expression=1)}
}
\arguments{
\item{gene_set}{the gene set, as a character vector.}

\item{counts}{a matrix or data frame of counts, or an object from which counts can be extracted. The dimension corresponding to genes must have dimnames, which will be matched to the gene identifiers in \code{gene_sets}.}

\item{counts_genes_by}{the dimension of \code{counts} corresponding to genes. Can be "rows" or "columns" or partial matches.}

\item{remove_missing_genes}{logical, whether to remove genes from the gene sets if the genes are not present in the counts object. Defaults to TRUE.}

\item{min_genes}{an integer, the minimum number of genes from a gene set that must be in the data. Gene sets with fewer genes overlapping with the dataset are dropped.}

\item{min_median_gene_set_expression}{numeric, the minimum value for median expression of genes in a gene set. Any gene sets with median expression lower than this value are removed.}
}
\value{
A character vector containing the genes that pass the criteria.  May be a zero-length vector, or identical to the input \code{gene_set} object.
}
\description{
This function filters a gene set based on a set of criteria. The standard criterion is median gene expression above a certain threshold, but other functionality may be built in at a later time. It can optionally remove genes not present in the counts object, and/or remove genes from gene sets if their median counts are below a certain threshold.
}
\seealso{
\code{\link{filter_gene_sets}}
}
