% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_median_zscore.R
\name{gene_set_median_zscore}
\alias{gene_set_median_zscore}
\title{Determine median z-scores of a set of genes in a counts object}
\usage{
\code{gene_set_median_zscore(gene_set, counts,
  counts_genes_by="rows",
  remove_missing_genes=TRUE,
  remove_low_count_genes=TRUE,
  ...)}
}
\arguments{
\item{gene_set}{the gene set, as a character vector.}

\item{counts}{a matrix or data frame of counts. The dimension corresponding to genes must have dimnames, which will be matched to the gene identifiers in \code{gene_sets}.}

\item{counts_genes_by}{the dimension of \code{counts} corresponding to genes. Can be "rows" or "columns" or partial matches.}

\item{remove_missing_genes}{logical, whether to remove genes from the gene sets if the genes are not present in the counts object. Defaults to TRUE.}

\item{remove_low_count_genes}{logical, whether to remove genes from the gene sets if the median expression level is below a threshold.  Implemented by a call to \code{gene_set_expressed}.}

\item{...}{optional parameters passed to \code{gene_set_expressed}. Ignored if \code{remove_low_count_genes} is set to FALSE.}
}
\value{
A numeric vector containing the median values, with one element for each library in \code{counts}.
}
\description{
This function determines the patient-level median z-score of a set of genes. Counts of each gene are first converted to z-scores, then the median value for each patient is calculated.
}

