% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_sets_incMatrix.R
\name{gene_sets_incMatrix}
\alias{gene_sets_incMatrix}
\title{Generate a binary incidence matrix for gene sets and counts}
\usage{
\code{gene_sets_incMatrix(
  gene_sets, counts,
  counts_genes_by="rows",
  gene_sets_by="rows",
  min_genes=2,
  method="binary")}
}
\arguments{
\item{gene_sets}{the gene sets. Either a list of gene sets, with each as a character vector, or a matrix, with one gene set per column, and the gene set names as column names. Names of list elements are used as gene set names.}

\item{counts}{a matrix or data frame of counts. The dimension corresponding to genes must have dimnames, which will be matched to the gene identifiers in \code{gene_sets}.}

\item{counts_genes_by}{the dimension of \code{counts} corresponding to genes. Can be "rows" or "columns" or partial matches.}

\item{gene_sets_by}{the dimension of the result that should correspond to gene sets. Can be "rows" or "columns" or partial matches. Not used if method is "index".}

\item{min_genes}{an integer, the minimum number of genes from a gene set that must be in the data. Gene sets with fewer genes overlapping with the dataset are dropped.}

\item{method}{character, the type of incidence object to return. Either "binary" for a binary matrix or "index" for a list of vector-based indices.}
}
\description{
This function creates a binary or index-based gene set incidence matrix. It expects gene sets as named lists of character vectors.
}

