% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetExactTest.R
\name{geneSetExactTest}
\alias{geneSetExactTest}
\title{Calculate Fisher's exact test statistics for a gene set in a subset of a gene list}
\usage{
\code{geneSetExactTest(gene_set, gene_list, index, alternative="greater", ...)}
}
\arguments{
\item{gene_set}{a character vector, the names of the genes in the gene set.}

\item{gene_list}{a character vector, the full gene list, including the subset to test.}

\item{index}{a logical or numeric vector. Either a logical vector with same length as \code{gene_list}, with TRUE/FALSE for each gene (where TRUE indicates membership in the subset), or a numeric vector with the indices from \code{gene_list} of the genes in the subset.}

\item{alternative}{the direction to test for enrichment. Passed to \code{fisher.test}. The default value, "greater", tests for members of \code{gene_set} found more often than expected in the specified subset of \code{gene_list}.}

\item{...}{optional arguments to be passed to \code{fisher.test}}
}
\value{
A data frame with a single row, containing the p-value, the upper and lower confidence intervals, and the estimated odds ratio.
}
\description{
Calculate Fisher's exact test statistics for a gene set in a subset of a list of genes, versus the
background to that subset (the full list of genes). i.e. test if the gene set is enriched in some
subset of a full gene list. The subset could be a WGCNA modules, a set of genes with significant
differential expression, etc. This tests against the null hypothesis of even distribution of the
contingency table; the test statistic is compared to a hypergeometric distribution. By default, it
tests for positive enrichment (genes in the gene set found more often in the subset than expected).
}

