% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetListExactTest.R
\name{geneSetListExactTest}
\alias{geneSetListExactTest}
\title{Calculate Fisher's exact test statistics for a number of gene sets in a subset of a gene list}
\usage{
\code{geneSetListExactTest(gene_sets, gene_list, index,
  alternative="greater", fdr.method="BH", ...)}
}
\arguments{
\item{gene_sets}{a list of character vectors, each vector containing the names of the genes in a single gene set.}

\item{gene_list}{a character vector, the full gene list, including the subset to test.}

\item{index}{a logical or numeric vector. Either a logical vector with same length as \code{gene_list}, with TRUE/FALSE for each gene (where TRUE indicates membership in the subset), or a numeric vector with the indices from \code{gene_list} of the genes in the subset.}

\item{alternative}{the direction to test for enrichment. Passed to \code{fisher.test} via \code{geneSetExactTest}. The default value, "greater", tests for members of each gene set found more often than expected in the specified subset of \code{gene_list}.}

\item{fdr.method}{the FDR calculation method, passed to \code{p.adjust}. The default, "BH", specifies the Benjamini-Hochberg approach.}

\item{...}{optional arguments to be passed to \code{fisher.test} via \code{geneSetExactTest}.}
}
\value{
A data frame with a row for each gene set, containing the name of the gene set, the p-value, the upper and lower confidence intervals, and the estimated odds ratio.
}
\description{
Calculate Fisher's exact test statistics for a list of gene sets in a subset of a list of genes,
versus the background to that subset (the full list of genes). i.e. test if each gene set is enriched
in some subset of a full gene list. It is a wrapper for geneSetExactTest, with multiple gene sets.
The subset could be a WGCNA modules, a set of genes with significant
differential expression, etc. This tests against the null hypothesis of even distribution of the
contingency table; the test statistic is compared to a hypergeometric distribution. By default, it
tests for positive enrichment (genes in the gene set found more often in the subset than expected).
}

