% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcgse_process.R
\name{pcgse_process}
\alias{pcgse_process}
\title{Reformat output of PCGSE for easier viewing}
\usage{
\code{pcgse_process(result, sort_by_PC=1, gene_sets=NULL)}
}
\arguments{
\item{result}{the output from PCGSE. Usually a list with two matrics: p.values and statistics.}

\item{sort_by_PC}{the number of the PC to sort results by}

\item{gene_sets}{the gene sets with which PCGSE was run. Used for pulling in the gene set names if they are missing}
}
\value{
A data frame, with one row for each gene set, and 3 columns for each PC tested by PCGSE. Columns include the test statistic, the p-value, and the adjsuted p-value.
}
\description{
This function reformats the output of PCGSE, allowing for easier viewing of the results. It calculates the
adjusted p-values, and converts the list into a data frame with the gene set, test statistics, p-values, and
adjusted p-values as columns.
}

