% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmt.R
\name{read_gmt}
\alias{read_gmt}
\title{Read a gene set in .gmt file format}
\usage{
\code{read_gmt(filename, out_format="matrix", empty_val=NA, name_loc=1, genes_start=3)}
}
\arguments{
\item{filename}{path to the .gmt file}

\item{out_format}{the format to output, either "list" or "matrix".}

\item{empty_val}{the value to use for empty cells. Ignored if out_format == "list".}

\item{name_loc}{index (position) of the gene set names in each gene set. List elements are assigned these as names. Defaults to 1.}

\item{genes_start}{index (position) of the first gene name in each gene set. Allows skipping descriptions and extra elements. Defaults to 3.}
}
\description{
This function reads a gene set in .gmt file format into a list. By default, it uses the first
element of each gene set as the name, and assumes the first gene in each set is the third element.
}

