% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WGCNAmodulesGeneSetListExactTest.R
\name{WGCNAmodulesGeneSetListExactTest}
\alias{WGCNAmodulesGeneSetListExactTest}
\title{Calculate Fisher's exact test statistics for a number of gene sets in WGCNA modules}
\usage{
\code{WGCNAmodulesGeneSetListExactTest(gene_sets, gene_modules, threshold=0.75,
  alternative="greater", fdr.method="BH", p.adjust_set="all",
  return_format="df", return_sort="p.value", ...)}
}
\arguments{
\item{gene_sets}{a list of character vectors, each vector containing the names of the genes in a single gene set.}

\item{gene_modules}{a data frame containing gene names ("gene"), WGCNA modules ("color_assigned"), and connectivity to the assigned module ("kME_color_assigned"). It should contain all genes included in the module construction, as the full set in \code{gene_modules} is used as the background.}

\item{threshold}{a scalar, the value of kME to use as a threshold for inclusion of genes in WGCNA modules. Genes with connectivity below this value are considered to NOT be members of the modules; they are included in the background list. Based on some minimal testing, threshold does not seem to have a huge effect on which gene sets are near the top of the list, but it does affect how many gene sets are found significant.}

\item{alternative}{the direction to test for enrichment. Passed to \code{fisher.test}. The default value, "greater", tests for members of each gene set found more often than expected in each WGCNA module.}

\item{fdr.method}{the FDR calculation method, passed to \code{p.adjust}. The default, "BH", specifies the Benjamini-Hochberg approach.}

\item{p.adjust_set}{the set of p-values to include in FDR q-value (adjusted p-value) calculation. The default, "all", aggregates all tests of all gene sets against all WGCNA modules, and calculates adjusted p-values for that set. Other possible values: "module" adjusts p-values for all gene sets tested within each module. Other values result in a warning, and adjustment as for "module".}

\item{return_format}{the format in which to return the results. The default, "df", returns a data frame, with a column for module. "list" returns a list, with each element containing the results for a single module. Other values result in an error.}

\item{return_sort}{the name(s) of the column(s) to sort results on. Defaults to "p.value".}

\item{...}{optional arguments to be passed to \code{fisher.test}.}
}
\value{
Either a data frame with a row for each combination of module and gene set; or a list of data frames, one for module, with each containing rows for the results of each gene set tested. In either case, the data frame(s) contain(s) the name of the gene set, the p-value, the upper and lower confidence intervals, and the estimated odds ratio.
}
\description{
Calculate Fisher's exact test statistics for a list of gene sets in WGCNA modules, versus the 
background of the full list of genes that the WGCNA modules were constructed from. i.e. test if 
each gene set is enriched in each module. It is a wrapper for geneSetListExactTest, with multiple 
subsets of the full gene list. This tests against the null hypothesis of even distribution of the
contingency table; the test statistic is compared to a hypergeometric distribution. By default, it
tests for positive enrichment (genes in the gene set found more often in the subset than expected).
}

