% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_cls.R
\name{write_cls}
\alias{write_cls}
\title{Write out phenotypes to a .cls file compatible with GSEA}
\usage{
\code{write_cls(file_prefix, counts, design, var_to_test, libID_col="lib.id")}
}
\arguments{
\item{file_prefix}{A string, the (optional) path and file name to output. The name of the variable, and ".cls" will be appended to this.}

\item{counts}{A matrix or data frame, the counts to be used in GSEA, with samples in columns and genes in rows. Should have columns labeled with sample identifiers. Necessary in order to determine the proper order of sample phenotype values in the output file.}

\item{design}{A data frame containing (at minimum) columns for the sample identifiers and phenotype values.}

\item{var_to_test}{A string, the name of the column in \code{design} that contains the phenotype values. Also added to the output file name.}

\item{libID_col}{Optional, the name or number of the column in \code{design} that contains sample identifiers matching the column names in \code{counts}. If "row.names" is specified, the rownames of \code{design} are used.}
}
\description{
Write out phenotype values to a .cls file, for use with the Java implementation of GSEA. Samples are
output in the order they are found in an associated counts object so that they are properly associated
in GSEA.
}

