% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_GSEA_counts.R
\name{write_GSEA_counts}
\alias{write_GSEA_counts}
\title{Write out counts to a .txt file compatible with GSEA}
\usage{
\code{write_GSEA_counts(filename, counts)}
}
\arguments{
\item{filename}{A string, the name of the file to output}

\item{counts}{A matrix or data frame of counts, or an object from which counts can be extracted. Should have samples in columns and genes in rows, and rows labeled with gene identifiers.}
}
\description{
Write out expression counts to a .txt file, for use with the Java implementation of GSEA. This function
formats the output to match the .txt counts file format used by GSEA.
}
