% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gene_sets.R
\name{filter_gene_sets}
\alias{filter_gene_sets}
\title{Filter gene sets}
\usage{
\code{filter_gene_sets(
gene_sets, counts,
counts_genes_by="rows",
min_genes=2,
min_median_expression=1,
remove_missing_genes=FALSE}
}
\arguments{
\item{gene_sets}{the gene sets. Either a list of gene sets, with each as a character vector, or a matrix or data frame, with one gene set per column, and the gene set names as column names. Names of list elements are used as gene set names.}

\item{counts}{a matrix or data frame of counts. The dimension corresponding to genes must have dimnames, which will be matched to the gene identifiers in \code{gene_sets}.}

\item{counts_genes_by}{the dimension of \code{counts} corresponding to genes. Can be "rows" or "columns" or partial matches.}

\item{min_genes}{an integer, the minimum number of genes from a gene set that must be in the data. Gene sets with fewer genes overlapping with the dataset are dropped.}

\item{min_median_expression}{numeric, the minimum value for median expression of genes in a gene set. Any gene sets with median expression lower than this value are removed.}

\item{remove_missing_genes}{logical, whether to remove genes from the gene sets if they are not found in counts. Defaults to FALSE.}
}
\description{
This function filters a gene set based on a set of criteria. The standard criterion is median gene expression above a certain threshold, but other functionality may be built in at a later time.
}

